Get rid of temporary value copies when doing a.b.c
Bending over backwards to support string.data here.
I added a special type to deal with the fact that
there is no actual storage for the *u8 when but we
do a ir.StructMemberPtr to try to get a **u8 for the
data member.

There are other options, for example we can look at
two instructions at once when we get an ir.StructMemberPtr,
but this approach allows assignment to .data to be caught
by the type checker.